+++ packages/material-ui-lab/src/BusyButton/BusyButton.test.js
@@ -1,9 +1,9 @@
 import * as React from 'react';
 import { getClasses } from '@material-ui/core/test-utils';
 import createMount from 'test/utils/createMount';
-import describeConformance from '../test-utils/describeConformance';
+import describeConformance from '@material-ui/core/test-utils/describeConformance';
 import BusyButton from './BusyButton';
-import Button from '../Button';
+import Button from '@material-ui/core/Button';
 
 describe('<BusyButton />', () => {
   const mount = createMount();
