+++ packages/material-ui/src/styles/props.d.ts
@@ -5,7 +5,6 @@ import { BadgeProps } from '../Badge';
 import { BottomNavigationActionProps } from '../BottomNavigationAction';
 import { BottomNavigationProps } from '../BottomNavigation';
 import { BreadcrumbsProps } from '../Breadcrumbs';
-import { BusyButtonProps } from '../BusyButton';
 import { ButtonBaseProps } from '../ButtonBase';
 import { ButtonGroupProps } from '../ButtonGroup';
 import { ButtonProps } from '../Button';
@@ -113,7 +112,6 @@ export interface ComponentsPropsList {
   MuiBottomNavigation: BottomNavigationProps;
   MuiBottomNavigationAction: BottomNavigationActionProps;
   MuiBreadcrumbs: BreadcrumbsProps;
-  MuiBusyButton: BusyButtonProps;
   MuiButton: ButtonProps;
   MuiButtonBase: ButtonBaseProps;
   MuiButtonGroup: ButtonGroupProps;
